Lots of changes.
Added children_func as a parameter, removed type_check as a parameter for each attribute and replaced it with attribute_check (which'll checks all parameters all at once [necessary for the url BBcode]), made it so __ can be validated and made it so that decompiled BBcodes return the original text and not some reparsed version of it.

Also, since request_var() passes stuff through htmlspecialchars, the regex used to identify characters was rewritten to look for &quot; and not ".